send numcopies stats when the count is 0
This won't send a 0 count for a number of copies which is higher than
the highest number of copies we currently have, because there are a lot
of those. Unfortunately, it would be desirable to do so if we have ever
had a number of copies that high in the past.

For example, if we currently have 1 file with 9 copies, and none with
any higher number, then we won't send a count of 0 for 10
copies. However, if we had 10 copies of a file in the past, then that
will be recorded in our statistics, and we're going to start getting the
wrong values in our charts now, because the charts will assume that we
still do. Hopefully it won't ever become a big problem.