if a given uuid is registered repeatedly, only list it once per email
When email address is changed, the old email may be present in the file
still, and will still be listed. This is ok, it just results in an extra
client page being made for that email. What we don't want is multiple
occurrances of the same uuid on a given client page.