simplify the download code so that it only checks for available disk space
It might ultimately be nice to check for other conditions so that we
come up for air more frequently, but if so then we want to take the
time to do it right so that users aren't confused. In particular, it
shouldn't say that it's "done" with a shard unless it actually is.