new scripts for creating and updating shards, along with some tests
These new scripts require fewer manual steps (no tsv files needed),
but admittedly cannot split a collection over multiple shards. On the
other hand, they can also record all information recieved from IA and
then play that back on subsequent runs. This allows for reproducable
tests, and it saves bandwith when you need to rerun the script.

The most important thing to note is that I'm adding a branch to every
shard called "shardmeta". This will contain a script that prints out a
list of items that ought to be in the shard. Currently this script
simply queries IA using the original collection list, allowing the
shards to grow as the collections on IA grow. We can revise these
scripts later if or when the collections grow large enough to be split
across shards.

add-shardmeta-branch.sh creates a new shardmeta branch in an existing
shard.

make-shard.sh creates a brand new shard out of a list of collections,
including the shardmeta branch.

update-shard.sh updates an existing shard with new items and metadata
from IA.

The test directory contains a script that tests make-shard.sh (and
thus update-shard.sh as well) on a pre-recorded (and slightly edited)
data set. Make sure you keep this up to date if you change the
scripts, etc. We should also write tests for the other scripts as we
go.