only update fsck.log weekly
The cron job runs more often than weekly, in order to find a time when
the disk is plugged in, computer is on, etc.

But, we only expire clients after 1 month of no fscks. So, actually
updating the fsck.log daily is excessive. Causes git-annex branch to
grow, and causes load on the server to sync it.

This uses a timestamp to only run the fast fsck if it's been a week since the
last one. Otherwise, the cron job avoids doing anything.